* use 0 instead of NULL * remove `state` local in favor of `switch(input()) {}`

+++ examples/snake/src/game.h
@@ -68,4 +68,4 @@ typedef struct {
 extern Snake snake;
 extern Point fruit;
 extern gfx_context vctx;
-extern int controller_mode;
+extern uint8_t controller_mode;
